# docs/specs/swagger.yaml
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/api/v1/fresh-products/advertise/{id}": {
      "get": {
        "tags": [
          "advertise-controller"
        ],
        "summary": "findById",
        "operationId": "findByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Advertise"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "advertise-controller"
        ],
        "summary": "update",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "advertiseDto",
            "description": "advertiseDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AdvertiseRequestDTO"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "advertise-controller"
        ],
        "summary": "delete",
        "operationId": "deleteUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/v1/fresh-products/buyer/{id}": {
      "get": {
        "tags": [
          "buyer-controller"
        ],
        "summary": "findById",
        "operationId": "findByIdUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BuyerResponseDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "buyer-controller"
        ],
        "summary": "update",
        "operationId": "updateUsingPUT_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "newBuyer",
            "description": "newBuyer",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BuyerRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "buyer-controller"
        ],
        "summary": "delete",
        "operationId": "deleteUsingDELETE_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/v1/fresh-products/product/{id}": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "summary": "findById",
        "operationId": "findByIdUsingGET_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProductResponseDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "product-controller"
        ],
        "summary": "update",
        "operationId": "updateUsingPUT_2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "newObj",
            "description": "newObj",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "product-controller"
        ],
        "summary": "delete",
        "operationId": "deleteUsingDELETE_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/v1/fresh-products/representative/{id}": {
      "get": {
        "tags": [
          "representative-controller"
        ],
        "summary": "findById",
        "operationId": "findByIdUsingGET_3",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RepresentativeResponseDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "representative-controller"
        ],
        "summary": "update",
        "operationId": "updateUsingPUT_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "newObj",
            "description": "newObj",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RepresentativeRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "representative-controller"
        ],
        "summary": "delete",
        "operationId": "deleteUsingDELETE_3",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/fresh-products/section": {
      "get": {
        "tags": [
          "section-controller"
        ],
        "summary": "findAll",
        "operationId": "findAllUsingGET_4",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SectionResponseDTO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "section-controller"
        ],
        "summary": "insert",
        "operationId": "insertUsingPOST_4",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "sectionRequestDTO",
            "description": "sectionRequestDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SectionRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SectionResponseDTO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/fresh-products/section/{id}": {
      "get": {
        "tags": [
          "section-controller"
        ],
        "summary": "findById",
        "operationId": "findByIdUsingGET_4",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SectionResponseDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "section-controller"
        ],
        "summary": "update",
        "operationId": "updateUsingPUT_4",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "sectionDTO",
            "description": "sectionDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SectionRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SectionResponseDTO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "section-controller"
        ],
        "summary": "delete",
        "operationId": "deleteUsingDELETE_4",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/v1/fresh-products/seller/{id}": {
      "get": {
        "tags": [
          "seller-controller"
        ],
        "summary": "findById",
        "operationId": "findByIdUsingGET_5",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SellerResponseDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "seller-controller"
        ],
        "summary": "update",
        "operationId": "updateUsingPUT_5",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "newObj",
            "description": "newObj",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SellerRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "seller-controller"
        ],
        "summary": "delete",
        "operationId": "deleteUsingDELETE_5",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/v1/fresh-products/warehouse/{id}": {
      "get": {
        "tags": [
          "warehouse-controller"
        ],
        "summary": "findById",
        "operationId": "findByIdUsingGET_6",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Warehouse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "warehouse-controller"
        ],
        "summary": "update",
        "operationId": "updateUsingPUT_6",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "nWarehouse1",
            "description": "nWarehouse1",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WarehouseRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "warehouse-controller"
        ],
        "summary": "delete",
        "operationId": "deleteUsingDELETE_6",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Advertise": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "free_shipping": {
          "type": "boolean"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "price": {
          "type": "number"
        },
        "product": {
          "$ref": "#/definitions/Product"
        },
        "seller": {
          "$ref": "#/definitions/Seller"
        },
        "status": {
          "type": "string",
          "enum": [
            "ATIVO",
            "FINALIZADO",
            "INATIVO",
            "PAUSADO",
            "PENDENTE",
            "REMOVIDO"
          ]
        }
      },
      "title": "Advertise"
    },
    "AdvertiseRequestDTO": {
      "type": "object",
      "required": [
        "price",
        "status"
      ],
      "properties": {
        "description": {
          "type": "string",
          "minLength": 4,
          "maxLength": 20
        },
        "free_shipping": {
          "type": "boolean"
        },
        "price": {
          "type": "number"
        },
        "product_id": {
          "type": "integer",
          "format": "int64"
        },
        "seller_id": {
          "type": "integer",
          "format": "int64"
        },
        "status": {
          "type": "string",
          "enum": [
            "ATIVO",
            "FINALIZADO",
            "INATIVO",
            "PAUSADO",
            "PENDENTE",
            "REMOVIDO"
          ]
        }
      },
      "title": "AdvertiseRequestDTO"
    },
...
    "BuyerRequestDTO": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string",
          "minLength": 2,
          "maxLength": 20
        },
        "pass": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      },
      "title": "BuyerRequestDTO"
    },
    "BuyerResponseDTO": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      },
      "title": "BuyerResponseDTO"
    },
...
    "Product": {
      "type": "object",
      "properties": {
        "category_refrigeration": {
          "type": "string",
          "enum": [
            "COLD",
            "FRESH",
            "FROZEN"
          ]
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "max_temperature": {
          "type": "number",
          "format": "float"
        },
        "min_temperature": {
          "type": "number",
          "format": "float"
        },
        "name": {
          "type": "string"
        }
      },
      "title": "Product"
    },
...
    "ProductRequestDTO": {
      "type": "object",
      "required": [
        "category_refrigeration",
        "max_temperature",
        "min_temperature",
        "name"
      ],
      "properties": {
        "category_refrigeration": {
          "type": "string",
          "enum": [
            "COLD",
            "FRESH",
            "FROZEN"
          ]
        },
        "description": {
          "type": "string",
          "minLength": 4,
          "maxLength": 20
        },
        "max_temperature": {
          "type": "number",
          "format": "float"
        },
        "min_temperature": {
          "type": "number",
          "format": "float"
        },
        "name": {
          "type": "string",
          "minLength": 4,
          "maxLength": 20
        }
      },
      "title": "ProductRequestDTO"
    },
    "ProductResponseDTO": {
      "type": "object",
      "properties": {
        "category_refrigeration": {
          "type": "string",
          "enum": [
            "COLD",
            "FRESH",
            "FROZEN"
          ]
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      },
      "title": "ProductResponseDTO"
    },
    "RepresentativeRequestDTO": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "job": {
          "type": "string",
          "enum": [
            "LIDER",
            "SUPERVISOR"
          ]
        },
        "name": {
          "type": "string"
        },
        "pass": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "warehouse_id": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "RepresentativeRequestDTO"
    },
    "RepresentativeResponseDTO": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "job": {
          "type": "string",
          "enum": [
            "LIDER",
            "SUPERVISOR"
          ]
        },
        "name": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "warehouse_id": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "RepresentativeResponseDTO"
    },
    "SectionRequestDTO": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "current_stock": {
          "type": "integer",
          "format": "int32",
          "minimum": 1,
          "maximum": 100,
          "exclusiveMinimum": false,
          "exclusiveMaximum": false
        },
        "id_warehouse": {
          "type": "integer",
          "format": "int64"
        },
        "max_teperature": {
          "type": "number",
          "format": "float",
          "minimum": -20,
          "maximum": 15,
          "exclusiveMinimum": false,
          "exclusiveMaximum": false
        },
        "min_teperature": {
          "type": "number",
          "format": "float",
          "minimum": -20,
          "maximum": 15,
          "exclusiveMinimum": false,
          "exclusiveMaximum": false
        },
        "name": {
          "type": "string",
          "minLength": 5,
          "maxLength": 15
        },
        "stock_limit": {
          "type": "integer",
          "format": "int32",
          "minimum": 1,
          "maximum": 100,
          "exclusiveMinimum": false,
          "exclusiveMaximum": false
        },
        "type": {
          "type": "string",
          "enum": [
            "COLD",
            "FRESH",
            "FROZEN"
          ]
        }
      },
      "title": "SectionRequestDTO"
    },
...
    "SectionResponseDTO": {
      "type": "object",
      "properties": {
        "current_stock": {
          "type": "integer",
          "format": "int32"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "max_teperature": {
          "type": "number",
          "format": "float"
        },
        "min_teperature": {
          "type": "number",
          "format": "float"
        },
        "name": {
          "type": "string"
        },
        "refrigeration_type": {
          "type": "string",
          "enum": [
            "COLD",
            "FRESH",
            "FROZEN"
          ]
        },
        "stock_limit": {
          "type": "integer",
          "format": "int32"
        },
        "warehouse_code": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "SectionResponseDTO"
    },
...
    "Seller": {
      "type": "object",
      "properties": {
        "advertise_list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Advertise"
          }
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "profile": {
          "type": "array",
          "uniqueItems": true,
          "items": {
            "type": "string",
            "enum": [
              "ADMIN",
              "BUYER",
              "REPRESENTATIVE",
              "SELLER"
            ]
          }
        },
        "profiles": {
          "type": "array",
          "uniqueItems": true,
          "items": {
            "type": "string",
            "enum": [
              "ADMIN",
              "BUYER",
              "REPRESENTATIVE",
              "SELLER"
            ]
          }
        },
        "username": {
          "type": "string"
        }
      },
      "title": "Seller"
    },
    "SellerRequestDTO": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "pass": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      },
      "title": "SellerRequestDTO"
    },
    "SellerResponseDTO": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      },
      "title": "SellerResponseDTO"
    },
...
    "Warehouse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "location": {
          "type": "string"
        },
        "nome": {
          "type": "string"
        }
      },
      "title": "Warehouse"
    },
    "WarehouseRequestDTO": {
      "type": "object",
      "properties": {
        "location": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "title": "WarehouseRequestDTO"
    },
...
  }
}